Added Error Message if Testing is Run too
